(refactor): migrate all tests and test helper code to TS
- woot, all TS now!
  - get that big blue line for language TS on GitHub haha

- for the most part, not much more to do other than .js -> .ts
  - for the utils had to add some types around

- also use ESM instead of CJS in tests and test helpers
  - require -> import
  - module.exports -> export
    - the indentation change here with utils/fixture made git treat
      it as a delete + new instead of a rename
