+++ test/e2e/tsdx-lint.test.ts
@@ -1,6 +1,6 @@
-const shell = require('shelljs');
+import * as shell from 'shelljs';
 
-const util = require('../utils/fixture');
+import * as util from '../utils/fixture';
 
 shell.config.silent = true;
