+++ test/integration/tsdx-build-withConfig.test.ts
@@ -1,8 +1,8 @@
-const shell = require('shelljs');
-const fs = require('fs-extra');
+import * as fs from 'fs-extra';
+import * as shell from 'shelljs';
 
-const util = require('../utils/fixture');
-const { execWithCache } = require('../utils/shell');
+import * as util from '../utils/fixture';
+import { execWithCache } from '../utils/shell';
 
 shell.config.silent = false;
